Add license info (#43)
* fixed capitalization

* Relabeling Create and Edit Event Pages (#22)

* Relabeled everything and added a few ts variables

* halfway made changes

* changed over the variable to a method

* changes method names

* formatted

* Back Buttons Labelling Correct (#20)

* altered button so the assignments and exams label was correct

* formatted

* got rid of if statement

* switched button placement (#25)

* Hide random question selection (#27)

* fixed capitalization

* Relabeling Create and Edit Event Pages (#22)

* Relabeled everything and added a few ts variables

* halfway made changes

* changed over the variable to a method

* changes method names

* formatted

* Back Buttons Labelling Correct (#20)

* altered button so the assignments and exams label was correct

* formatted

* got rid of if statement

* switched button placement (#25)

* removed add questions from html

* removed related methods and imports from ts

* added cancel button for editing/creating challenges

---------


* Relabel complete to view (#29)

* fixed capitalization

* Relabeling Create and Edit Event Pages (#22)

* Relabeled everything and added a few ts variables

* halfway made changes

* changed over the variable to a method

* changes method names

* formatted

* Back Buttons Labelling Correct (#20)

* altered button so the assignments and exams label was correct

* formatted

* got rid of if statement

* switched button placement (#25)

* changed complete to be view for students

---------


* Move category below question settings (#31)

* fixed capitalization

* Relabeling Create and Edit Event Pages (#22)

* Relabeled everything and added a few ts variables

* halfway made changes

* changed over the variable to a method

* changes method names

* formatted

* Back Buttons Labelling Correct (#20)

* altered button so the assignments and exams label was correct

* formatted

* got rid of if statement

* switched button placement (#25)

* moved nav under header

---------


* Hide see details mcq (#33)

* fixed capitalization

* Relabeling Create and Edit Event Pages (#22)

* Relabeled everything and added a few ts variables

* halfway made changes

* changed over the variable to a method

* changes method names

* formatted

* Back Buttons Labelling Correct (#20)

* altered button so the assignments and exams label was correct

* formatted

* got rid of if statement

* switched button placement (#25)

* added condition to hide details for mcq

---------


* initial license html

* minor improvements

* version 1

* version 2

* added is

* wip

* wip

* improvements to island

* info for teachers only

* improved tool-tip

* improved tool-tip

* add_instructor_label_to_courses (#47)

* Hide random question selection (#27)

* fixed capitalization

* Relabeling Create and Edit Event Pages (#22)

* Relabeled everything and added a few ts variables

* halfway made changes

* changed over the variable to a method

* changes method names

* formatted

* Back Buttons Labelling Correct (#20)

* altered button so the assignments and exams label was correct

* formatted

* got rid of if statement

* switched button placement (#25)

* removed add questions from html

* removed related methods and imports from ts

* added cancel button for editing/creating challenges

---------


* Relabel complete to view (#29)

* fixed capitalization

* Relabeling Create and Edit Event Pages (#22)

* Relabeled everything and added a few ts variables

* halfway made changes

* changed over the variable to a method

* changes method names

* formatted

* Back Buttons Labelling Correct (#20)

* altered button so the assignments and exams label was correct

* formatted

* got rid of if statement

* switched button placement (#25)

* changed complete to be view for students

---------


* Move category below question settings (#31)

* fixed capitalization

* Relabeling Create and Edit Event Pages (#22)

* Relabeled everything and added a few ts variables

* halfway made changes

* changed over the variable to a method

* changes method names

* formatted

* Back Buttons Labelling Correct (#20)

* altered button so the assignments and exams label was correct

* formatted

* got rid of if statement

* switched button placement (#25)

* moved nav under header

---------


* Hide see details mcq (#33)

* fixed capitalization

* Relabeling Create and Edit Event Pages (#22)

* Relabeled everything and added a few ts variables

* halfway made changes

* changed over the variable to a method

* changes method names

* formatted

* Back Buttons Labelling Correct (#20)

* altered button so the assignments and exams label was correct

* formatted

* got rid of if statement

* switched button placement (#25)

* added condition to hide details for mcq

---------


* added instructor label

* changed to tailwind margins and got rid of unnecessary status if statement

---------


* wrote NightModeService to use in header and footer

* made modifications

* removed providers

---------
