+++ src/app/components/footer/footer.component.ts
@@ -1,4 +1,5 @@
 import {Component} from '@angular/core'
+import {NightModeService} from "@app/_services/night-mode.service"
 
 @Component({
     selector: 'app-footer',
@@ -6,4 +7,10 @@ import {Component} from '@angular/core'
     styleUrls: ['./footer.component.scss']
 })
 export class FooterComponent {
+    constructor(private nightModeService: NightModeService) {
+    }
+
+    getNightMode(): boolean {
+        return this.nightModeService.getNightMode()
+    }
 }
