+++ src/app/course/course-homepage/course-homepage.component.html
@@ -13,17 +13,26 @@
                 <p class="tui-island__paragraph tui-text_body-xl flex-space">
                     Improve your skills by solving practice
                     questions.</p>
-                <div class="course-view-button tui-space_top-3">
-                    <a
-                        appearance="primary"
-                        [routerLink]="['/course', courseId, 'practice']"
-                        (click)="practiceClickLog()"
-                        class="tui-form__button"
-                        size="m"
-                        tuiButton
-                    >
-                        View
-                    </a>
+                <div class="flex items-end justify-between">
+                    <div class="course-view-button tui-space_top-3">
+                        <a
+                            appearance="primary"
+                            [routerLink]="['/course', courseId, 'practice']"
+                            (click)="practiceClickLog()"
+                            class="tui-form__button"
+                            size="m"
+                            tuiButton
+                        >
+                            View
+                        </a>
+                    </div>
+                    <tui-marker-icon *ngIf="user.is_teacher"
+                                     size="s"
+                                     src="tuiIconInfo"
+                                     tuiHintMode="onDark"
+                                     tuiPointerHint="All of our questions are licensed under CC BY-NC-SA 4.0"
+                                     tuiHintDirection="top-middle"
+                    ></tui-marker-icon>
                 </div>
             </div>
         </tui-island>
